backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-cloud-vision==3.4.4
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
gunicorn==20.1.0
pymongo==4.5.0
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
